Split GitHub Projects engine helpers
