+++ cli/python/base_github_projects/tests/test_subprocess_timeouts.py
@@ -5,6 +5,8 @@ import subprocess
 import pytest
 
 from base_github_projects import engine
+from base_github_projects import project_git
+from base_github_projects import project_graphql
 
 
 def test_infer_repo_from_git_passes_timeout(monkeypatch: pytest.MonkeyPatch) -> None:
@@ -17,10 +19,10 @@ def test_infer_repo_from_git_passes_timeout(monkeypatch: pytest.MonkeyPatch) -> 
 
     def fake_run(command: list[str], **kwargs: object) -> subprocess.CompletedProcess[str]:
         assert command == ["git", "config", "--get", "remote.origin.url"]
-        assert kwargs["timeout"] == engine.GIT_COMMAND_TIMEOUT_SECONDS
+        assert kwargs["timeout"] == project_git.GIT_COMMAND_TIMEOUT_SECONDS
         return completed
 
-    monkeypatch.setattr(engine.subprocess, "run", fake_run)
+    monkeypatch.setattr(project_git.subprocess, "run", fake_run)
 
     assert engine.infer_repo_from_git() == "basefoundry/base"
 
@@ -29,7 +31,7 @@ def test_infer_repo_from_git_returns_none_on_timeout(monkeypatch: pytest.MonkeyP
     def fake_run(command: list[str], **kwargs: object) -> subprocess.CompletedProcess[str]:
         raise subprocess.TimeoutExpired(command, kwargs["timeout"])
 
-    monkeypatch.setattr(engine.subprocess, "run", fake_run)
+    monkeypatch.setattr(project_git.subprocess, "run", fake_run)
 
     assert engine.infer_repo_from_git() is None
 
@@ -44,10 +46,10 @@ def test_run_graphql_passes_timeout(monkeypatch: pytest.MonkeyPatch) -> None:
 
     def fake_run(command: list[str], **kwargs: object) -> subprocess.CompletedProcess[str]:
         assert command == ["gh", "api", "graphql", "--input", "-"]
-        assert kwargs["timeout"] == engine.GITHUB_GRAPHQL_TIMEOUT_SECONDS
+        assert kwargs["timeout"] == project_graphql.GITHUB_GRAPHQL_TIMEOUT_SECONDS
         return completed
 
-    monkeypatch.setattr(engine.subprocess, "run", fake_run)
+    monkeypatch.setattr(project_graphql.subprocess, "run", fake_run)
 
     assert engine.run_graphql("query Viewer { viewer { login } }", {}) == {
         "data": {"viewer": {"login": "codeforester"}}
@@ -58,7 +60,7 @@ def test_run_graphql_reports_timeout(monkeypatch: pytest.MonkeyPatch) -> None:
     def fake_run(command: list[str], **kwargs: object) -> subprocess.CompletedProcess[str]:
         raise subprocess.TimeoutExpired(command, kwargs["timeout"])
 
-    monkeypatch.setattr(engine.subprocess, "run", fake_run)
+    monkeypatch.setattr(project_graphql.subprocess, "run", fake_run)
 
     with pytest.raises(engine.ProjectError) as excinfo:
         engine.run_graphql("query Viewer { viewer { login } }", {})
